+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
+- #18 Added `sample_shipped` status and `deliver` transition
+- #18 Added Courier and Couriers list
 - #11 Additional `send` transition in AR/Sample workflow
 - #7 New field "Study code" in Client (instead of Tax Number)
 - #2 New fields in Analysis Request (and visible in AR Add)